MaskFormer/code/requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1742562417367/work
antlr4-python3-runtime==4.9.3
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1736248061654/work
audioread==3.0.1
black==25.1.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
Cartopy==0.24.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi==1.16.0
cftime @ file:///home/conda/feedstock_root/build_artifacts/cftime_1725400453617/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
click==8.1.8
cloudpickle==3.1.1
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1731428322366/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
Cython==3.0.12
decorator==5.1.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@400a49c1ec11a18dd25aea3910507bc3bcd15794
filelock==3.13.1
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1739279199749/work
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1743732186120/work
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1733299680490/work
google-pasta @ file:///home/conda/feedstock_root/build_artifacts/google-pasta_1733852499742/work
grpcio @ file:///home/conda/feedstock_root/build_artifacts/grpc-split_1740784437750/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py @ file:///home/conda/feedstock_root/build_artifacts/h5py_1739952247544/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
huggingface-hub==0.30.2
hydra-core==1.3.2
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.37.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
iopath==0.1.9
Jinja2==3.1.4
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1741854587245/work
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
lazy_loader==0.4
libclang==18.1.1
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1710435156458/work
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1733250460757/work
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
ml-dtypes @ file:///home/conda/feedstock_root/build_artifacts/ml_dtypes_1725475027258/work
mpmath==1.3.0
msgpack==1.0.8
munkres==1.1.4
mypy-extensions==1.0.0
namex @ file:///home/conda/feedstock_root/build_artifacts/namex_1733858070416/work
netCDF4 @ file:///home/conda/feedstock_root/build_artifacts/netcdf4_1733253191879/work
networkx==3.3
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1707225380409/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=51131fd8fc130cd168aecaf1bc0ea85f92e8ffebf211772ceb16ac2e7f10d7ca
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.11.0.86
opt_einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1733687912731/work
optree @ file:///home/conda/feedstock_root/build_artifacts/optree_1744034198035/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.2.0
pooch==1.8.1
portalocker==3.1.1
protobuf @ file:///home/conda/feedstock_root/build_artifacts/protobuf_1731365437114/work/bazel-bin/python/dist/protobuf-5.28.3-cp310-abi3-linux_x86_64.whl#sha256=1a22bd966bf84a4610fd6a1624465b3b8c18b8fa7a0834dcd09b387a3888cd3f
pycocotools==2.0.8
pycparser==2.21
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1743089729650/work
pyproj==3.7.1
pyshp==2.3.1
PySide6==6.9.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz==2025.2
PyYAML==6.0.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1743371105/work/dist
safetensors==0.5.3
scikit-image==0.25.2
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1739790642651/work/dist/scipy-1.15.2-cp310-cp310-linux_x86_64.whl#sha256=9e52bad6c3294d1a5b04a13632118ca2157130603c6c018c2d710162b223b27e
shapely==2.1.0
shiboken6==6.9.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
soundfile==0.12.1
soxr==0.3.7
sympy==1.13.1
tabulate==0.9.0
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1733820672324/work/tensorboard-2.18.0-py3-none-any.whl#sha256=107ca4821745f73e2aefa02c50ff70a9b694f39f790b11e6f682f7d326745eab
tensorboard_data_server @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-data-server_1728639716925/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=a9e810807f394a98b62699603d2070bd51f04159295d8a158fb17145b7ff0564
tensorflow @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-split_1739089378357/work/tensorflow_pkg/tensorflow-2.18.0-cp310-cp310-linux_x86_64.whl#sha256=2e41f4f6c1b6a1dd961f6e2af4eb99fce4e87895b7ac0a8ddb0527ee3a00533b
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_estimator @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-split_1739089378357/work/tensorflow-estimator/wheel_dir/tensorflow_estimator-2.15.0-py2.py3-none-any.whl#sha256=b28747f5a37fbc1a432d7c963bc057ab06d5ea2f757fd891d1a63d975dd95d58
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1743676967121/work
tifffile==2025.5.10
timm==1.0.15
tomli==2.2.1
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615898999/work
tqdm==4.67.1
triton==2.1.0
typing_extensions==4.12.2
tzdata==2025.2
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1736692496989/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1744323578849/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
wrapt @ file:///home/conda/feedstock_root/build_artifacts/wrapt_1736869474897/work
xarray==2025.4.0
yacs==0.1.8
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard==0.23.0
